Fix uint usages, which is not defined on Windows.

+++ libcore/include/sirikata/core/network/SSTImpl.hpp
@@ -361,7 +361,7 @@ private:
 
   std::tr1::weak_ptr<Connection<EndPointType> > mWeakThis;
 
-  uint mNumInitialRetransmissionAttempts;
+  uint16 mNumInitialRetransmissionAttempts;
 
   google::protobuf::LogSilencer logSilencer;
